Game_round now function as it should
The game round loop is functioning and working, the loop is stopping at five points. Good decsion to merge the scoreboard in this class to save time at the moment. That is now on the nice to have list along with incorporating the functions from rules engine which have the wording for each weapons win, using beat at the moment. Errors still with the @name instance variable updating. Refactoring...